Worked on spectrogram error plot and on confidence intervals for perceptual quality figure.

+++ wav_resample.py
@@ -29,4 +29,4 @@ wavfile.write(
     Fs,
     np.array(
         signal,
-        dtype=np.int16))
+        dtype=np.float))
